requirements.txt
tensorflow
torch
torchvision
keras
Pillow